% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entry.R
\name{disableProxy}
\alias{disableProxy}
\title{Disable proxy by removing environment variables}
\usage{
disableProxy()
}
\description{
Disable proxy by removing environment variables
}
\examples{
disableProxy()
}
